requirements.txt
streamlit
sentence-transformers
fpdf2
requests
Pillow
easyocr
opencv-python-headless
python-dotenv
numpy
pytesseract

